types: migrate Int core to safe integers

+++ crates/seseragi-lowering/src/core.rs
@@ -292,7 +292,7 @@ pub enum CoreExpr {
     Unit {
         origin: SourceSpan,
     },
-    Int64 {
+    Integer {
         value: String,
         origin: SourceSpan,
     },
